requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
config==0.5.0.post0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-registration==2.4.1
gunicorn==20.0.4
Pillow==8.0.0
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.1
soupsieve==2.1
sqlparse==0.4.1
whitenoise==5.2.0
